refactor(cli): extract the line-mode REPL from agent.rs into repl.rs
`agent.rs` carried two unrelated responsibilities: `abi agent` subcommand
dispatch (plan/multi/train/spawn/browser/os) and the whole non-TTY REPL —
help text, model-id validation, session state, ten slash handlers, and the
stdin loop. At 942 lines it was the last open row on the cleanup goal.

The REPL moves verbatim to `crates/abi-cli/src/repl.rs`, along with the
eight tests that only exercise it. `agent_tui_line_mode` becomes
`repl::line_mode()`; nothing else was renamed and no logic changed.

agent.rs 942 -> 520 lines (off the hotspot watch list); repl.rs is 442.

Pure refactor. Verified nothing outside agent.rs referenced any moved item
before the move, and no golden fixture pins REPL text. Full ./tools/check.sh
green, goldens byte-identical, and `abi agent tui` fed /status, /model
fable-5, /profile, /help, /quit emits the same bytes as before.
